AirPlay: Fix a streaming video/audio stopping after 1 minute
For some reason, Apple TV closes an inactive HTTP connection that is
used to play a video/audio and stops playing after 60 seconds. TCP
keep-alive packets do not help. This patch adds an explicit HTTP
keep-alive request that is sent via AirPlay every 50 seconds to maintain
the connection and keep the media playing.

Closes https://github.com/ConnectSDK/Connect-SDK-iOS/issues/106