Fix a crash when discovering LG Speaker device
The device description XML lists only one `<device>` tag under the
`<deviceList>` section, which is parsed to return a dictionary, whereas
the code expected to get an array. This resulted in sending an
unrecognized selector `-[__NSDictionaryM enumerateObjectsUsingBlock:]`.
The patch explicitly checks for the returned type and adds the
dictionary to an array if necessary.